Refactor security config and fix database initialization
