+++ ecommerce-backend/src/main/java/com/ecommerce/backend/entity/RefreshToken.java
@@ -11,7 +11,7 @@ import java.time.LocalDateTime;
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
-@SuperBuilder
+@Builder
 public class RefreshToken extends BaseEntity {
 
     @Column(name = "token", nullable = false, unique = true, length = 500)
