+++ ecommerce-backend/src/main/java/com/ecommerce/backend/entity/Review.java
@@ -9,7 +9,7 @@ import lombok.*;
 @Setter
 @NoArgsConstructor
 @AllArgsConstructor
-@SuperBuilder
+@Builder
 public class Review extends BaseEntity {
 
     @ManyToOne(fetch = FetchType.LAZY)
