+++ ecommerce-backend/src/main/resources/application.yml
@@ -123,6 +123,12 @@ spring:
     hibernate:
       ddl-auto: create-drop
     show-sql: true
+    defer-datasource-initialization: true
+    
+  sql:
+    init:
+      mode: always
+      data-locations: classpath:data-h2.sql
     
 logging:
   level:
